refactor: rename *_service -> *
