+++ books/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class BooksServiceConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "books_service"
+    name = "books"
