+++ books/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path, include
 from rest_framework import routers
 
-from books_service.views import BookViewSet
+from books.views import BookViewSet
 
 app_name = "books"
