+++ books/views.py
@@ -1,8 +1,8 @@
 from rest_framework import viewsets
 
-from books_service.models import Book
-from books_service.permissions import IsAdminOrReadOnly
-from books_service.serializers import BookSerializer
+from books.models import Book
+from books.permissions import IsAdminOrReadOnly
+from books.serializers import BookSerializer
 
 
 class BookViewSet(viewsets.ModelViewSet):
