+++ borrowings/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class BorrowingsServiceConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "borrowings_service"
+    name = "borrowings"
