+++ borrowings/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
 
-from borrowings_service.views import BorrowingViewSet
+from borrowings.views import BorrowingViewSet
 
 app_name = "borrowings"
