+++ notifications/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class TelegramServiceConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "telegram_service"
+    name = "notifications"
