+++ users/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class UsersServiceConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "users_service"
+    name = "users"
