+++ users/views.py
@@ -1,7 +1,7 @@
 from rest_framework import generics
 from rest_framework.permissions import IsAuthenticated
 
-from users_service.serializers import UserSerializer
+from users.serializers import UserSerializer
 
 
 class CreateUserView(generics.CreateAPIView):
